requirements.txt
fastapi==0.85.0
uvicorn==0.19.0
tensorflow==2.10.0
python-multipart
pillow
opencv-python-headless
numpy